==> 01_env_data/bin/09_geotraces_GP13_CTD.Rmd <==
---
title: "GP13 CTD tidier"
output:   
  github_document:
    toc: yes
    toc_depth: 2
    fig_width: 7
    fig_height: 5
    dev: png
    keep_html: yes
    html_preview: yes
---

This uses the GEOTRACES Intermediate Data Product 2017 (Version 2)

The license for this data does not allow it to be distributed outside of the original. See here:

> GEOTRACES Intermediate Data Product (IDP) Download Agreement
> The GEOTRACES programme is keen to ensure that the very significant effort and expertise involved in making trace-element and isotope measurements is acknowledged as fully as possibly in subsequent publications.
>
>Users of the GEOTRACES Intermediate Data Product are expected to abide to the following rules regarding citation
>
>To the greatest extent possible, please cite all relevant publications from researchers that made the measurements you use in your work. Details of publications that should be cited are provided point-by-point in the IDP dataset (in the ODV and ASCII versions) and will be updated on the online database as new papers are published. Where your research deals particularly with data measured by a single group of data originators, you are invited to please contact that group to discuss your work prior to publication and jointly consider the synergy and mutual benefit of co-authorship where appropriate.
>
>Where other constraints prevent citation of all relevant publications, for instance where there is a journal limitation on the maximum number of publications that can be cited, or if the dataset is only used in a minor supportive way, please cite the data compilation itself (as below). In such cases, also please cite any original individual papers that you rely on particularly heavily for your research and interpretations.
>
>Where using data from the IDP2017 product in your publication, please cite the data compilation as: Schlitzer, R., Anderson, R. F., Masferrer Dodas, E, et al., The GEOTRACES Intermediate Data Product 2017, Chem. Geol. (2018), https://doi.org/10.1016/j.chemgeo.2018.05.040 .
>
>Where using data from the IDP2014 product in your publication, please cite the data compilation as: Mawji, E., et al., The GEOTRACES Intermediate Data Product 2014, Mar. Chem. (2015), http://dx.doi.org/10.1016/j.marchem.2015.04.005 .
>
>Users of the GEOTRACES Intermediate Data Product shall not distribute downloaded data to any third party.

To follow along you muust register with the [the British Oceanographic Data Centre](https://www.bodc.ac.uk/geotraces/data/idp2017/) and download the data. We will be using the `CTD Sensor Data` in ODV format.

Note that there [are some issues](https://www.bodc.ac.uk/data/documents/nodb/544232/) with some of the IDP data. Fortunately, none of these issues apply to the samples corresponding to the metagenomes.

Also note that this data is not particularly tidy... What we really want out of it is the chlorophyll data in order to calculate where the DCM is in the water column because for some reason no one measured chlorophyll from the discrete bottles! Unfortunately, the chlorophyll data exists in either concentration transformed fluorescence in mg/m3 or in just raw beam flourescence. This means that the two are not directly comparable, for example in statistical models, (at least without the formula to convert between raw beam fluorescence and concentration). However, we can at least get an idea where the fluorescence peaks at the subsurface and this way we can calculate where the DCM is and then match that roughly to where are bottle samples were located.

## Preprocessing of the CTD data file using GNU unix tools
Split into individual cruise transects. Not sure if this is necessary but I thought that maybe columns are different between cruise sections. Also some sections join much nicer (GA02, GA03) whereas GP13 and GA10 are particularly problematic

```{bash}
grep "^Cruise" ../data/geotraces/GEOTRACES_IDP2017_v1_CTD_Sensor_Data/GEOTRACES_IDP2017_v1_CTD_Sensor_Data.odv_1c69fb81_1.txt > ../data/geotraces/GEOTRACES_IDP2017_v1_CTD_Sensor_Data/GP13_CTD.tsv

grep "^GP13" ../data/geotraces/GEOTRACES_IDP2017_v1_CTD_Sensor_Data/GEOTRACES_IDP2017_v1_CTD_Sensor_Data.odv_1c69fb81_1.txt >> ../data/geotraces/GEOTRACES_IDP2017_v1_CTD_Sensor_Data/GP13_CTD.tsv
```

## Processing with R
```{r message=FALSE, warning=FALSE}
library(here)
library(tidyverse)
library(janitor)
library(fuzzyjoin)
library(magrittr)
library(lubridate)
library(RcppRoll)
library(missRanger)
library(withr)

`%nin%` <- Negate(`%in%`)
```

## Read metagenomes metadata
```{r}
GP13.mg <- read_tsv(here::here("data", "geotraces", "geotraces_formatted_metadata.tsv")) %>% 
  filter(section=="GP13") %>% 
  select(sampleID, bottle_num_BODC, station_num_BODC, lat, lon, depth) %>%
  mutate(station_num_BODC=as.numeric(station_num_BODC)) %>%
  mutate(bottle_num_BODC=as.numeric(bottle_num_BODC)) %>%
  mutate(depthmin=depth-1.5,
         depthmax=depth+1.5) %>%
  mutate(depthmin=ifelse(depthmin < 0, 0, depthmin))
```


```{r message=FALSE, warning=FALSE}
GP13.CTD <- read_tsv(here::here("data", "geotraces", "GEOTRACES_IDP2017_v1_CTD_Sensor_Data", "GP13_CTD.tsv"), col_names = TRUE) %>%
  janitor::clean_names() %>%
  janitor::remove_empty(which = "cols") %>%
  select(-contains("qv_iode"), 
         -contains("infos_"), 
         -contains("standard_dev_"), 
         -qv_odv_sample) %>%
  filter(ctdprs_dbar < 300) %>%
  mutate(longitude_degrees_east=ifelse(longitude_degrees_east > 180, longitude_degrees_east-360, longitude_degrees_east)) %>%
  rename(lat=latitude_degrees_north, lon=longitude_degrees_east) %>%
  select(-type, -yyyy_mm_dd_thh_mm_ss_sss, -bot_depth_m, -processing_info, 
         -info_document, -operators_cruise_name, -cruise_report, -ship_name, 
         -chief_scientist, -geotraces_scientist) %>%
  drop_na(fluorescence_raw_data_arbitrary_units)%>%
  mutate(lon=round(lon, 3), lat=round(lat, 3))
```

```{r}
colnames(GP13.CTD)
```

## Format and normalize Chla
```{r}
GP13.CTD.chla <- GP13.CTD %>%
  group_by(lat, lon, station) %>%
  mutate(chla_moving_average = roll_mean(fluorescence_raw_data_arbitrary_units, 10, align="center", fill=NA)) %>%
  fill(chla_moving_average, .direction = "updown") %>%
  mutate(chla_norm = scale(chla_moving_average, center=T)) %>%
  mutate(offset = min(chla_norm)) %>%
  mutate(chla_norm = ifelse(offset < 0, chla_norm - offset, chla_norm)) %>%
  mutate(chla_norm = chla_norm/max(chla_norm)) %>%
  select(cruise, station, lon, lat, depth=ctdprs_dbar, chla_norm, offset) %>%
  ungroup() %>%
  mutate(station=paste("S", station, sep="_"))
```

## Get DCM depth for each station
```{r}
GP13.CTD.dcm_d <- GP13.CTD.chla %>%
  distinct(lat, lon, depth, .keep_all = TRUE) %>%
  mutate(simpleID = paste("ID", group_indices(., lon, lat), sep="_")) %>%
  group_by(cruise, station, lon, lat) %>%
  filter(chla_norm >= 0.9*max(chla_norm)) %>%
  mutate(dcm_depth=median(depth)) %>%
  ungroup() %>%
  select(cruise, station, lon, lat, dcm_depth) %>%
  distinct()
```


## Determine DCM depth
First determine the depth of the DCM based on the full CTD profile. Will save this information and add to final metadata.
```{r}
GP13.CTD.chla1 <- left_join(GP13.CTD, select(GP13.CTD.dcm_d, -station)) %>% 
  group_by(lat, lon, station) %>%
  mutate(chla_moving_average = roll_mean(fluorescence_raw_data_arbitrary_units, 10, align="center", fill=NA)) %>%
  fill(chla_moving_average, .direction = "updown") %>%
  mutate(chla_norm = scale(chla_moving_average, center=T)) %>%
  mutate(offset = min(chla_norm)) %>%
  mutate(chla_norm = ifelse(offset < 0, chla_norm - offset, chla_norm)) %>%
  mutate(chla_norm = chla_norm/max(chla_norm)) %>%
  select(cruise, station, lon, lat, depth=ctdprs_dbar, chla_norm, dcm_depth, offset) %>%
  ungroup()

GP13.CTD.dcm_d1 <- GP13.CTD.chla1 %>%
  filter(depth > 1) %>%
  group_by(lat, lon, station) %>%
  mutate(dcm.layer = ifelse(chla_norm > 0.5*max(chla_norm, na.rm=T), 1, 0)) %>%
  mutate(dcm.max = ifelse(chla_norm >= 0.9*max(chla_norm, na.rm=T), 1, 0)) %>%
  select(cruise, station, lon, lat, depth, chla_norm, dcm_depth, dcm.layer, dcm.max) %>%
  ungroup()
```

Plot of chlorophyll profile for each station and the depth of the DCM from the rolling average approach
```{r message=FALSE, warning=FALSE}
ggplot() +
  geom_path(data=GP13.CTD.chla1, aes(y=depth, x=chla_norm, color=chla_norm, group=interaction(station,lon, lat)), size=0.25) +
  scale_color_viridis_c(option="C") +
  geom_point(data=GP13.CTD.dcm_d1 %>% filter(dcm.layer==1), 
             aes(y=depth, x=chla_norm), color="red") +
  scale_y_reverse(limits=c(300,0)) +
  theme_bw()
```

Can look at stations individually to check (not run)
```{r eval=FALSE, include=TRUE}
ggplot() +
  geom_path(data=GP13.CTD.chla, aes(y=depth, x=chla_norm, color=chla_norm, group=interaction(station,lon, lat)), size=0.25) +
  scale_color_viridis_c(option="C") +
  geom_point(data=GP13.CTD.dcm %>% filter(dcm.layer==1), 
             aes(y=depth, x=chla_norm), color="red") +
  facet_wrap(~station) +
  scale_y_reverse(limits=c(300,0)) +
  theme_bw()
```

## Mapping
Check how well the bottle samples and the CTD casts line up.

For some reason there is no CTD data at three different bottle locations but I checked and there is discrete CTD bottle data at those geo coordinates.

```{r}
library(sf)
library("rnaturalearth")
library("rnaturalearthdata")

world <- ne_countries(scale = "medium", returnclass = "sf")

mg.loc <- GP13.mg %>% distinct(lat, lon)
ctd.loc <- GP13.CTD %>% distinct(lat, lon)

ggplot(data = world) + 
  geom_sf(fill = "grey70", color=NA) +
  geom_point(data=mg.loc, aes(x=lon, y=lat), color="red", size=3) +
  geom_point(data=ctd.loc, aes(x=lon, y=lat), color="blue") +
  coord_sf(xlim = c(140, 180), ylim = c(-45, -25), expand = FALSE) +
  theme_bw()
```

```{r}
ggplot(data = world) + 
  geom_sf(fill = "grey70", color=NA) +
  geom_point(data=mg.loc, aes(x=lon, y=lat), color="red", size=3) +
  geom_point(data=ctd.loc, aes(x=lon, y=lat), color="blue") +
  coord_sf(xlim = c(-180, -147), ylim = c(-45, -25), expand = FALSE) +
  theme_bw()
```

#Final Formatting
Convert coordinates from the CTD to metagenome coordinates
```{r}
GP13.mg.convert <- geo_left_join(GP13.mg, distinct(GP13.CTD.dcm_d1, lon, lat), 
                        by = c("lat","lon"),
                        distance_col="geojoin.dist.km",
                        max_dist = 62) %>%
  mutate(geojoin.dist.km=geojoin.dist.km*1.60934) %>%
  group_by(sampleID) %>%
  filter(geojoin.dist.km==min(geojoin.dist.km)) %>%
  ungroup() %>%
  select(sampleID, lon=lon.y, lat=lat.y, geojoin.dist.km, depth, depthmin, depthmax)
```

Join with the metagenomes
```{r}
GP13.dcm.a <- fuzzy_left_join(GP13.mg.convert, GP13.CTD.dcm_d1, 
                      by=c("lon" = "lon",
                           "lat" = "lat",
                        "depthmin" = "depth",
                        "depthmax" = "depth"),
                       match_fun = list(`==`, `==`, `<=`, `>=`)
                      ) %>%
  mutate(ddiff = abs(depth.x-depth.y)) %>%
  group_by(sampleID) %>%
  filter(ddiff == min(ddiff)) %>%
  select(sampleID, geojoin.dist.km, chla_norm, dcm_depth, dcm.layer, dcm.max) %>% 
  group_by(sampleID) %>%
  mutate(chla_norm=mean(chla_norm)) %>%
  filter(dcm.layer==max(dcm.layer)) %>%
  filter(dcm.max==max(dcm.max)) %>%
  distinct() %>%
  ungroup()

GP13.dcm.b <- GP13.mg %>% filter(sampleID %nin% pull(GP13.dcm.a, sampleID)) %>%
  mutate(chla_norm=NA, dcm.layer=NA, dcm.max=NA) %>%
  select(sampleID, chla_norm, dcm.layer, dcm.max) 

GP13.dcm.missing <- bind_rows(GP13.dcm.a, GP13.dcm.b) %>%
  left_join(select(GP13.mg, -depthmin, -depthmax), .)  %>%
  group_by(lat, lon) %>%
  fill(geojoin.dist.km, .direction="updown") %>%
  mutate(chla_norm = ifelse(is.na(chla_norm) & depth > 350, 0, chla_norm),
           dcm.layer = ifelse(is.na(dcm.layer) & depth > 350, 0, dcm.layer), 
           dcm.max = ifelse(is.na(dcm.max) & depth > 350, 0, dcm.max)) %>%
  ungroup()
```

# Imputation
Unfortunately, there are complete missing values in the Chla data for the last 19 samples. Here we'll use random forest imputation with the missRanger package.

The missRanger package uses the ranger package to do fast missing value imputation by chained random forest. As such, it serves as an alternative implementation of the beautiful 'MissForest' algorithm, see vignette.

```{r}
with_preserve_seed(GP13.dcm.imputed <- GP13.dcm.missing %>%
  select(lat, lon, depth, chla_norm, dcm_depth, dcm.layer, dcm.max)%>%
  missRanger(verbose = 1, num.trees = 500) %>%
  mutate(dcm.layer=round(dcm.layer, digits=0),
         dcm.max=round(dcm.max, digits=0),
         dcm_depth=round(dcm_depth, digits=1)))
```

```{r}
GP13.dcm.final <- left_join(select(GP13.dcm.missing, -chla_norm, -dcm.layer,
                                   -dcm.max, -dcm_depth), GP13.dcm.imputed) %>%
  arrange(station_num_BODC, depth) %>%
  arrange(sampleID)
```

write the results for later use
```{r}
write_tsv(GP13.dcm.final, here::here("data", "intermediate_files", "GP13_mg_ctd.tsv"))
```

```{r}
sessionInfo()
```
